Added a sneaky little conditional to player movement
If the player is playing in solo mode, is in the air, and is NOT
controlling the character anymore, they keep their current velocity.

The characters will essentially act as if the player WOULD have kept
holding that direction while in air. This should make certain levels
(level 4) easier-and-or-possible and should be intuitive enough to pick
up naturally.